Finalização do projeto Spring Boot com HAL

+++ src/main/java/br/senaigo/mobile/entities/Product.java
@@ -20,7 +20,7 @@ public class Product extends ResourceSupport {
 	@Id
 	@GeneratedValue (strategy = GenerationType.AUTO)
 	@Column(name="id")
-	private Integer idProduct;
+	private Long idProduct;
 	private String name;
 	private String description;
 	private BigDecimal value;
